requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
pytest
httpx
